Revert "Add Ice and Water Chest"
