--- Assets/LowPolyTropicalEnvironment_LITE/Prefabs/Chest/IceChest.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 90c851bec66c7544fa5512d5f23f358c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
