--- Assets/LowPolyTropicalEnvironment_LITE/Prefabs/Chest/InvisibilityChest.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 297d75b443ca77743b6f27e71fac66ca
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
