--- Assets/Prefabs/EnvironmentObjects/Geyser/ChestGeyser.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 90f4678f56f65054bbfc2cd1df9ac347
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
