--- Assets/Scripts/Player/Abilities/IceBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 83ee0a8d9321b3949abc95ddd1c1e1ee
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
