+++ Assets/Scripts/Player/PlayerMovementController.cs
@@ -1,21 +1,19 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using Unity.VisualScripting;
-
 //using UnityEditor.Experimental.GraphView;
 using UnityEngine;
-//using UnityEngine.UI; // Fï¿½r den Zugriff auf UI-Komponenten
+//using UnityEngine.UI; // Für den Zugriff auf UI-Komponenten
 using UnityEngine.Animations.Rigging;
 using UnityEngine.InputSystem;
 using UnityEngine.UIElements;
 
 public class PlayerMovementController : MonoBehaviour
 {
-    // Referenzen auf die UI-Images fï¿½r die Inventar-Slots
+    // Referenzen auf die UI-Images für die Inventar-Slots
     [SerializeField] private UnityEngine.UI.Image invSlot1Image;
 
-    // Sprites fï¿½r jede Waffe
+    // Sprites für jede Waffe
     [SerializeField] private Sprite gunSprite;
     [SerializeField] private Sprite swordSprite;
 
@@ -38,7 +36,6 @@ public class PlayerMovementController : MonoBehaviour
 
     // special sword attack
     public bool freezed = false;
-    IceChest iceChest;
 
     //Movement
     CharacterController _characterController;
@@ -148,7 +145,6 @@ public class PlayerMovementController : MonoBehaviour
         _sword = _swordTransform.GetComponent<SwordTemplate>();
         _barrelSpawnTransform = transform.Find("BarrelSpawn");
         _barrelSpawn = _barrelSpawnTransform.GetComponent<BarrelSpawner>();
-        if (GameObject.Find("IceChest")!= null) iceChest = GameObject.Find("IceChest").GetComponent<IceChest>();
 
 
         _playerControls.CharacterControls.Move.started += context => { Move(context); };
@@ -179,7 +175,7 @@ public class PlayerMovementController : MonoBehaviour
                 swordModelSwapper.enableRenderer();
                 AimRig.weight= 0f;
                 UpdateWeaponUI(swordSprite);
-                OnSwordSwitched?.Invoke(); // Lï¿½se das Event fï¿½r den Wechsel zum Schwert aus
+                OnSwordSwitched?.Invoke(); // Löse das Event für den Wechsel zum Schwert aus
             }
             else if (weapon == 1)
             {
@@ -188,7 +184,7 @@ public class PlayerMovementController : MonoBehaviour
                 swordModelSwapper.disableRenderer();
                 AimRig.weight = 1f;
                 UpdateWeaponUI(gunSprite);
-                OnGunSwitched?.Invoke(); // Lï¿½se das Event fï¿½r den Wechsel zur Gun aus
+                OnGunSwitched?.Invoke(); // Löse das Event für den Wechsel zur Gun aus
             }
         };
         _playerControls.CharacterControls.WaterCannon.started += context =>
@@ -209,8 +205,7 @@ public class PlayerMovementController : MonoBehaviour
         };
         _playerControls.CharacterControls.IceSword.started += context =>
         {   
-            if(iceChest== null) return;
-            if(weapon == 1 && iceChest.swordUnlocked) {
+            if(weapon == 1) {
                 bool _isStriking= true;
                 _sword.Attack(true);
                 if (_isStriking) { PlayerAnimator.SetTrigger("Attack"); }
