+++ Assets/Scripts/Weapon/SwordTemplate.cs
@@ -6,38 +6,20 @@ using UnityEngine;
 
 public class SwordTemplate : WeaponTemplate
 {
-
+    
     [SerializeField] private float swordAttackRange = 2f;
     [SerializeField] private float swordCooldown = 1.0f;
     [SerializeField] private int swordDamage = 5;
     [SerializeField] private int IceSwordDamage = 1;
     [SerializeField] float freezingTime = 10f;
     [SerializeField] Material freezingMaterial;
-    [SerializeField] private float energyMax = 100f; // Maximale Energiemenge des Ice Swords
-    public float currentEnergy; // Aktuelle Energiemenge des Ice Swords
-    private float energyCost = 5f;
-    IceChest iceChest;
-    public IceBar iceBar;
-    Transform parentTransform;
-
-    public void Start()
-    {
-        parentTransform = transform.parent;
-        if (parentTransform.CompareTag("Player"))
-        {
-            iceBar = FindObjectOfType<IceBar>();
-            currentEnergy = energyMax;
-            iceBar.SetMaxEnergy(energyMax);
-            iceChest = FindObjectOfType<IceChest>();
-        }
-    }
-
+    
 
     public float GetswordAttackRange()
     {
-        return swordAttackRange;
+        return swordAttackRange; 
     }
-
+    
 
     public bool Attack(bool freeze = false)
     {
@@ -50,26 +32,20 @@ public class SwordTemplate : WeaponTemplate
                 tag = hit[i].tag;
                 if (opposingFraction.Contains(tag))
                 {
-                    if (freeze && hit[i].gameObject.GetComponent<BasicEnemy>() != null && currentEnergy >= energyCost)
-                    {
+                    if(freeze && hit[i].gameObject.GetComponent<BasicEnemy>()!= null){
                         hit[i].gameObject.GetComponent<EnemyTemplate>().Hurt((int)(IceSwordDamage * (UnityEngine.Random.value > GameController.instance.CritChance ? 1 : 1 + GameController.instance.CritDmg)));
                         StartCoroutine(hit[i].gameObject.GetComponent<BasicEnemy>().freeze(freezingTime, freezingMaterial));
-                        currentEnergy -= energyCost;
-                        iceBar.SetEnergy(currentEnergy);
                     }
-                    else if (freeze && hit[i].gameObject.GetComponent<EvilChest>() != null && currentEnergy >= energyCost)
-                    {
+                    else if(freeze && hit[i].gameObject.GetComponent<EvilChest>()!= null){
                         hit[i].gameObject.GetComponent<EnemyTemplate>().Hurt((int)(IceSwordDamage * (UnityEngine.Random.value > GameController.instance.CritChance ? 1 : 1 + GameController.instance.CritDmg)));
                         StartCoroutine(hit[i].gameObject.GetComponent<EvilChest>().freeze(freezingTime, freezingMaterial));
-                        currentEnergy -= energyCost;
-                        iceBar.SetEnergy(currentEnergy);
                     }
 
-                    else if (hit[i].gameObject.GetComponent<EnemyTemplate>() != null)
+                    else if(hit[i].gameObject.GetComponent<EnemyTemplate>()!= null)
                     {
                         hit[i].gameObject.GetComponent<EnemyTemplate>().Hurt((int)(swordDamage * (UnityEngine.Random.value > GameController.instance.CritChance ? 1 : 1 + GameController.instance.CritDmg)));
                     }
-                    else if (hit[i].gameObject.GetComponent<Health>() != null)
+                    else if(hit[i].gameObject.GetComponent<Health>()!= null)
                     {
                         hit[i].gameObject.GetComponent<Health>().TakeDamage(swordDamage);
                     }
@@ -81,13 +57,12 @@ public class SwordTemplate : WeaponTemplate
 
 
 
-                }
-                else if (hit[i].GetComponent<Chest>())
-                {
+                } 
+                else if (hit[i].GetComponent<Chest>()) {
                     Chest chest = hit[i].GetComponent<Chest>();
                     chest.openChest();
                 }
-
+                
             }
 
             onCooldown = true;
@@ -96,20 +71,11 @@ public class SwordTemplate : WeaponTemplate
             return true;
         }
         return false;
-
+        
     }
 
+    
+    
 
-    public IEnumerator ChargeEnergy(float amount)
-    {
-        while (iceChest.isPlayerInRange)
-        {
-            currentEnergy = Mathf.Min(currentEnergy + amount, energyMax);
-            iceBar.SetEnergy(currentEnergy);
-            yield return new WaitForSeconds(1f);
-        }
-
-    }
-
-
+    
 }
