--- Assets/Scripts/World/Chest/ChestWater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fe6cfb7021d0a8841a86e11b4e40ff3e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
