--- Assets/Scripts/World/Chest/IceChest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 759cd669962322542bbc73e9af5f7177
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
