--- Assets/Scripts/World/Chest/InvisibilityChest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6c828422309d68f44884be4189e24388
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
