--- Assets/Scripts/World/Chest/WaterChest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 61b2bd55db52bfd4786e4e50b86b3be1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
