+++ Assets/Souns/chestOpen.mp3.meta
